// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  auth: typeof auth;
  http: typeof http;
  showtimes: typeof showtimes;
  voters: typeof voters;
}>;
// ...
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;